+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # master
 
-* Add `config.view_component.preview_route` to set the endpoint for component previews. By default `/rails/view_component` is used.
+* Add `config.view_component.preview_route` to set the endpoint for component previews. By default `/rails/view_components` is used.
 
     *Juan Manuel Ramallo*