+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # master
 
+# 2.6.0
+
 * Add `config.view_component.preview_route` to set the endpoint for component previews. By default `/rails/view_components` is used.
 
     *Juan Manuel Ramallo*